+++ shared/teams/common/index.tsx
@@ -5,6 +5,7 @@ export {default as ChannelsWidget} from './channels-widget'
 export {useChannelMeta, useAllChannelMetas} from './channel-hooks'
 export {default as SelectionPopup} from './selection-popup'
 export {default as EnableContactsPopup} from './enable-contacts'
+export {default as useTeamLinkPopup} from './use-team-link-popup'
 
 export const usePhoneNumberList = () => {
   const [phoneNumbers, setPhoneNumbers] = React.useState([{key: 0, phoneNumber: '', valid: false}])
