+++ shared/teams/external-team.tsx
@@ -5,6 +5,7 @@ import * as Container from '../util/container'
 import * as Constants from '../constants/teams'
 import * as RPCGen from '../constants/types/rpc-gen'
 import * as Chat2Gen from '../actions/chat2-gen'
+import {useTeamLinkPopup} from './common'
 import {pluralize} from '../util/string'
 import {memoize} from '../util/memoize'
 import capitalize from 'lodash/capitalize'
@@ -146,40 +147,7 @@ const Header = ({info}: ExternalTeamProps) => {
       })
     )
 
-  const shareURLApp = `keybase://team-page/${teamname}`
-  const shareURLWeb = `https://keybase.io/team/${teamname}`
-
-  const {popupAnchor, setShowingPopup, popup} = Kb.usePopup(getAttachmentRef => {
-    const content = (
-      <Kb.Box2 direction="vertical" style={styles.linkPopupContainer} gap="small" fullWidth={true}>
-        <Kb.Text type="Header">Share a link to this team</Kb.Text>
-        <Kb.Box2 direction="vertical" gap="tiny" alignSelf="stretch" alignItems="stretch">
-          <Kb.Text type="Body">In the Keybase app:</Kb.Text>
-          <Kb.CopyText text={shareURLApp} shareSheet={true} />
-        </Kb.Box2>
-        <Kb.Box2 direction="vertical" gap="tiny" alignSelf="stretch" alignItems="stretch">
-          <Kb.Text type="Body">On the web:</Kb.Text>
-          <Kb.CopyText text={shareURLWeb} shareSheet={true} />
-        </Kb.Box2>
-        {Styles.isMobile && (
-          <Kb.Button type="Dim" label="Close" fullWidth={true} onClick={() => setShowingPopup(false)} />
-        )}
-      </Kb.Box2>
-    )
-    if (Styles.isMobile) {
-      return <Kb.MobilePopup>{content}</Kb.MobilePopup>
-    }
-    return (
-      <Kb.Overlay
-        position="bottom left"
-        style={styles.overlay}
-        attachTo={getAttachmentRef}
-        onHidden={() => setShowingPopup(false)}
-      >
-        {content}
-      </Kb.Overlay>
-    )
-  })
+  const {popupAnchor, setShowingPopup, popup} = useTeamLinkPopup(teamname || '')
 
   const metaInfo = (
     <Kb.Box2 direction="vertical" alignSelf="stretch" gap={Styles.isMobile ? 'small' : 'tiny'}>
@@ -282,9 +250,6 @@ const styles = Styles.styleSheetCreate(() => ({
   headerContainer: {
     ...Styles.padding(0, Styles.globalMargins.small),
   },
-  linkPopupContainer: {
-    ...Styles.padding(Styles.globalMargins.small, Styles.globalMargins.tiny),
-  },
   memberBody: {
     flex: 1,
     paddingRight: Styles.globalMargins.tiny,
@@ -293,7 +258,6 @@ const styles = Styles.styleSheetCreate(() => ({
     marginLeft: Styles.globalMargins.xtiny,
     marginRight: Styles.globalMargins.xtiny,
   },
-  overlay: {backgroundColor: Styles.globalColors.white, marginTop: Styles.globalMargins.tiny},
   tabs: {
     backgroundColor: Styles.globalColors.white,
     width: '100%',
